+++ runtime/shared/IntlPunctuation.js
@@ -23,14 +23,15 @@
  *  Note: Please keep this in sync with www/html/js/mobile/lib/intl-core.js.
  */
 
-const FbtHooks = require('FbtHooks');
-const IntlPhonologicalRewrites = require('IntlPhonologicalRewrites');
+import FbtHooks from 'FbtHooks';
+import IntlPhonologicalRewrites from 'IntlPhonologicalRewrites';
+import IntlRedundantStops from 'IntlRedundantStops';
 
 /**
  * Regular expression snippet containing all the characters that we
  * count as sentence-final punctuation.
  */
-const PUNCT_CHAR_CLASS = ('[.!?' +
+export const PUNCT_CHAR_CLASS = ('[.!?' +
   '\u3002' + // Chinese/Japanese period
   '\uFF01' + // Fullwidth exclamation point
   '\uFF1F' + // Fullwidth question mark
@@ -41,41 +42,6 @@ const PUNCT_CHAR_CLASS = ('[.!?' +
   '\u0E2F' + // Thai ellipsis
   '\uFF0E' + // Fullwidth full stop
   ']': string);
-
-const ENDS_IN_PUNCT_REGEXP = new RegExp(
-  PUNCT_CHAR_CLASS +
-    '[)"\'' +
-    // JavaScript doesn't support Unicode character
-    // properties in regexes, so we have to list
-    // all of these individually. This is an
-    // abbreviated list of the "final punctuation"
-    // and "close punctuation" Unicode codepoints,
-    // excluding symbols we're unlikely to ever
-    // see (mathematical notation, etc.)
-    '\u00BB' + // Double angle quote
-    '\u0F3B' + // Tibetan close quote
-    '\u0F3D' + // Tibetan right paren
-    '\u2019' + // Right single quote
-    '\u201D' + // Right double quote
-    '\u203A' + // Single right angle quote
-    '\u3009' + // Right angle bracket
-    '\u300B' + // Right double angle bracket
-    '\u300D' + // Right corner bracket
-    '\u300F' + // Right hollow corner bracket
-    '\u3011' + // Right lenticular bracket
-    '\u3015' + // Right tortoise shell bracket
-    '\u3017' + // Right hollow lenticular bracket
-    '\u3019' + // Right hollow tortoise shell
-    '\u301B' + // Right hollow square bracket
-    '\u301E' + // Double prime quote
-    '\u301F' + // Low double prime quote
-    '\uFD3F' + // Ornate right parenthesis
-    '\uFF07' + // Fullwidth apostrophe
-    '\uFF09' + // Fullwidth right parenthesis
-    '\uFF3D' + // Fullwidth right square bracket
-    '\\s' +
-    ']*$',
-);
 
 type Rules = $ReadOnlyArray<[RegExp, (string => string) | string]>;
 
@@ -125,7 +91,7 @@ function _getRules(locale: ?string): Rules {
  *        "{name}(y)i...")
  * Returns: String with phonological rules applied (e.g., "Ozguri...")
  */
-function applyPhonologicalRules(text: string): string {
+export function applyPhonologicalRules(text: string): string {
   const rules = _getMemoizedRules(FbtHooks.getViewerContext().locale);
   let result = text;
 
@@ -139,20 +105,34 @@ function applyPhonologicalRules(text: string): string {
 }
 
 /**
- * Checks whether a string ends in sentence-final punctuation. This logic is
- * about the same as the PHP ends_in_punct() function; it takes into account
- * the fact that we consider a string like "foo." to end with a period even
- * though there's a quote mark afterward.
+ * Map all equivalencies to the normalized key for the stop category.  These
+ * are the entries in the redundancy mapping
  */
-function endsInPunct(str: string): boolean {
-  if (typeof str !== 'string') {
-    return false;
+const _normalizedStops = new Map();
+for (const norm in IntlRedundantStops.equivalencies) {
+  for (const eq of [norm].concat(IntlRedundantStops.equivalencies[norm])) {
+    _normalizedStops.set(eq, norm);
   }
-  return ENDS_IN_PUNCT_REGEXP.test(str);
 }
 
-module.exports = {
-  PUNCT_CHAR_CLASS,
-  endsInPunct,
-  applyPhonologicalRules,
-};
+const _redundancies = new Map();
+for (const prefix in IntlRedundantStops.redundancies) {
+  _redundancies.set(prefix, new Set(IntlRedundantStops.redundancies[prefix]));
+}
+
+function isRedundant(rawPrefix: string, rawSuffix: string): boolean {
+  const prefix = _normalizedStops.get(rawPrefix);
+  const suffix = _normalizedStops.get(rawSuffix);
+  return _redundancies.get(prefix)?.has(suffix) === true;
+}
+
+/**
+ * If suffix is redundant with prefix (as determined by the redundancy map),
+ * return the empty string, otherwise return suffix.
+ */
+export function dedupeStops(prefix: string, suffix: string): string {
+  // We can naively grab the last "character" (a general Unicode "no-no") from
+  // our string because we know our set of stops we test against have no
+  // diacritics nor lie outside the BMP
+  return isRedundant(prefix[prefix.length - 1], suffix) ? '' : suffix;
+}